bug creando index V-0.0.3

+++ app/src/main/java/com/nazdesigns/polascope/USoT/FBCaller.java
@@ -185,7 +185,7 @@ public abstract class FBCaller {
                 for ( DataSnapshot obj : dataSnapshot.getChildren() ){
                     switch (obj.getKey()){
                         case "index":
-                            otherIndex = (double) obj.getValue();
+                            otherIndex = (Long) obj.getValue();
                             break;
                         case "raiz":
                             parentfbId = (String) obj.getValue();
